Utilize rails 5 attributes API, soft convert YAML serialization to JSON
The regular `serialize` approach raises errors in rails 5 if a query searches
for a string on a serialized-hash-field, e.g.:

`Project.where(title_translations: '%foo%')`

With the use of the new attributes API we can avoid these kinds of errors by
treating the serialized field as a string column with some curry, implemented
in `SerialTranslator::TranslationType`.

It also allows us to migrate away from YAML serializations to JSON without a
full migration and possible downtime. The use of YAML serialization might be
deprecated in the future.